Simulated Clock Support (#66)
This implements some core functionality to allow a simulated clock to be driven based on the timestamps of inbound messages. This is a building block towards deterministic replay of messages.

When the simulated clock is enabled, timer callbacks will also be invoked with the correct simulated time interval in between received messages.

Support for invoking timers on nodes that do not have any subscribers will need to be added at a future time along with support for a universal simulated clock and not just a node-specific one.