Disable service tests for now (#6)
These tests pass locally but fail often (but not always) in the CI
environment. They depend on multiple threads running and TCP sockets
connecting, so there's plenty of system-level interactions that get in
the way of determinism. We'll keep them disabled for the time being
until these issues are sorted out.