Subscriber Event Loop Refactor (#32)
Before this PR, subscribers would call back with new messages on the
same underlying message receive threads (eCAL uses one thread per
receiver).

The message message consumer would use the subscriber callback to push
to the FIFO and then pull from the FIFO in the event loop thread to call
back to the application. The FIFO itself was used as the data structure
that was shared between the threads.

Now, the handoff from the receive thread to the event loop thread
happens in the subscribers themselves. This is better for a few reasons.
Now the subscriber deserializes to a dynamically allocated protobuf that
is then handed off to the event loop efficiently via a unique_ptr. This
has the benefit of no longer needing a mutex to protect the FIFO. It
also makes direct usage of the subscriber from applications more
friendly. And now because the subscriber is calling back on the event
loop, the message consumer doesn't need to post any work to the event
loop and can just issue callbacks directly.