[health] Don't update health when only description changes (#48)
Proposal to not update health when only the description changes. I think
this encourages users to put as much info as possible in the
description, with the understanding that the description will only be
logged on real transitions between states.

Otherwise users are encouraged to keep health descriptions consistent
across multiple updates by removing debug information, or worse
implement their own update logic which checks if they have a significant
update (change to state or status code) before calling the trellis
update function.