[inbox] Add loopback receiver (#70)
Add a loopback receiver type to the inbox. This is very helpful for apps
that want to be stateless, as it allows them to maintain state but only
if it is also published (hence their state could be reconstructed
offboard for introspection, like simulation or replay). It also means
the app can be a simpler function of the inputs, rather than pulling
some information from state and some from the inbox. The app pays very
little for this, only the serialization/send cost, as we maintain the
loopback type in the same format as it was handed to us. An example
"stateless" app that might use this is as follows:

```
struct State{
  int count;
};

struct ToProto{
  StateProto operator()(const State& state);
};

// Main logic is stateless function.
std::optional<State> Run(const auto& inputs){
  const auto& [state, input] = inputs;
  if (!state.has_value() || !input.has_value()) return std::nullopt;
  return {{.count = state->message.count + input->message.val()}};
}

struct App{
  Inbox<Loopback<State, StateProto, ToProto>, Latest<Input>> inbox;

  // App function can be rather simple, Inbox hides complexity.
  void Tick(const TimePoint& time) {
    // Not const to allow move.
    auto output = Run(inbox.GetMessages(time));
    if (output.has_value()) inbox.Send(std::move(*output));
  }
};
```