[inbox] Add AllLatest receive type, a generalization of NLatest (#93)
Add an AllLatest receive type for the inbox, which is similar to NLatest
except that it is unbounded in how many elements it can return. This
makes it slightly less efficient, but is necessary for handling "the
last N seconds" of messages on a topic that publishes at an inconsistent
frequency.

The main efficiency losses are that we have to allocate memory
dynamically (not a big deal though, as we will grow to some max size and
likely stay there) and that messages are copied out of the subscriber's
memory pool and into the dynamic ring buffer (again not too big of an
issue, as large messages are mostly on the heap anyway).