[bazel] Switch to grpc py_proto_library rule (#104)
The current py_proto_library rule we use from protobuf breaks after
protobuf v3.20.3 and there are no newer rules which are compatible with
the syntax, so we must move to grpc's py_proto_library implementation,
which is different in that the py_proto_library doesn't take .proto
files as srcs, instead it accepts exactly one proto_library label in
deps, the py_proto_library inherits dependencies and sources from the
proto_library. This requires upgrading protobuf (v3.21.9 is the furthest
possible due to other interdependencies in the haystack repo,) and
adding grpc as a dependency.

As of this commit the trellis repo should be compatible with bazel
versions up to 7.3.1.