discovery: don't crash on update stats and reset handle on sub dtor (#168)
This attempts to resolve this error that occurred during destruction: 
```
what():  Attempt to retrive registered sample that doesn't exist. Handle = 9
```

The only times this could be called is either `GetSampleId` or
`UpdatePubSubStats`. Since this happened during destruction of a
subscriber, then its likely `UpdatePubSubStats` being called right after
the subscriber gets destroyed, but the timer callback was still in the
event loop.

The PR turns the error in `UpdatePubSubStats` from an exception to a
logged error. Failure to update stats shouldn't cause a crash