[AUT-5593] Add Conversion to Publisher (#175)
Add support for automatic conversion between serializable types
(protobufs) to other types (simple C++ structs).

This is first in a sequence of PRs that adds support for automatic
conversion between types that are convenient for serialization and types
that are convenient to work with.

This PR modifies the publisher and establishes the pattern that will
used to implement conversions for the subscriber, message consumer, and
inbox.

Conversion can be done with free functions or functors; conversion for
dynamic publishers is not supported (and is guarded against explicitly).
A unit test shows the basic call pattern.

Notably, this PR does not affect existing publishers because the
additional (template) arguments are optional.