[core][publisher] Dynamic publisher checks validity of schema
In the scenario where there is a dynamic publisher, one raw subscriber,
and one typed subscriber, the publisher needs the schema of the topic
from the typed subscriber. but if it gets the discovery sample from the
raw sub first, then it never gets the schema from the typed sub.

This PR checks for the validity of the schema in the sample and
re-registers if the current one is not valid.