Support for publishing node-level metrics starting with timer overruns (#200)
Adds a new feature where the node can be configured to publish metrics
at a specified interval. The first metric published is timer overruns.
Metrics publisher utility's dependency on Node was removed to break a
cyclic dependency when integrating into Node. This is a breaking change
to the metrics publisher APIs, so usage of MetricsPublisher will need to
be updated.